🚚 Move google-firebase-storage functions to a subdirectory
