+++ src/functions/google-firebase-storage/index.ts
@@ -1,0 +1,1 @@
+export { GoogleFirebaseStorageMergeRules } from './merge-rules.js';
