+++ src/functions/index.ts
@@ -12,7 +12,7 @@ import {
   EmulatorStopForSpanner,
 } from './emulator/index.js';
 import { GoogleAppCheckGenerateToken } from './google-app-check/index.js';
-import { GoogleFirebaseStorageMergeRules } from './google-firebase-storage-merge-rules.js';
+import { GoogleFirebaseStorageMergeRules } from './google-firebase-storage/index.js';
 import { GoogleFirestoreMergeRules } from './google-firestore-merge-rules.js';
 import { GoogleIdentityPlatformGenerateCustomToken } from './google-identity-platform-generate-custom-token.js';
 import { GoogleIdentityPlatformGenerateToken } from './google-identity-platform-generate-token.js';
