+++ src/functions/project-push-artefact-cloud-functions.spec.ts
@@ -14,7 +14,7 @@ import {
   EmulatorStartForFirebaseStorage,
   EmulatorStopForFirebaseStorage,
 } from './emulator/index.js';
-import { GoogleFirebaseStorageMergeRules } from './google-firebase-storage-merge-rules.js';
+import { GoogleFirebaseStorageMergeRules } from './google-firebase-storage/index.js';
 import { ProjectPushArtefactForCloudFunctions } from './project-push-artefact-cloud-functions.js';
 
 describe('ProjectPushArtefactForCloudFunctions', () => {
